S-53159 renamed gitHubEventsToApiResponse.js to commitEventsToApiResponse.js
